Namespace the creator, updater, validator, presenter and result
