--- lib/book_presenter.rb
@@ -1,21 +1,0 @@
-class BookPresenter
-  def initialize(book)
-    @book = book
-  end
-
-  def id
-    book.id
-  end
-
-  def title
-    book.title.titleize
-  end
-
-  def filename
-    book.filename
-  end
-
-  private
-
-  attr_reader :book
-end
