+++ lib/books/presenter.rb
@@ -1,0 +1,23 @@
+module Books
+  class Presenter
+    def initialize(book)
+      @book = book
+    end
+
+    def id
+      book.id
+    end
+
+    def title
+      book.title.titleize
+    end
+
+    def filename
+      book.filename
+    end
+
+    private
+
+    attr_reader :book
+  end
+end
